Save line count in Viewer
No need to pass around line count from Main to Viewer if we can get that
number from the text String parameter. Store this as part of the Viewer
structure to avoid passing references of it around.

This will make it easier to update the line count once files are being
edited.