Only print the characters that fit on the screen
Don't show all characters till the end of the char slice till. Characters
printed past the terminal's right edge go nowhere, so this isn't a big
problem. It is cleaner to avoid asking for them to be shown.
Plus it sets the limits in case the right edge will ever be used for
something.