Add jumping to a specific line
Action GoToLine will activate a mode in which numbers are read as digits
of a line number. These digits will be displayed in the status line.
Then when the Go action happens it will trigger the jump to that line.

If line number provided is past the end of the file, nothing will
happen.

This commit brought to you by flight ZB7988