Accept all keys in the config file
Use pattern matching to read the config file and connect any key with an
action. If it is only one char, use the char. Then there is a list of
keynames (Tab, Enter, Esc, ...) which are matched first by string length
and then by value. Also match the keycombos that start with "ctrl+" and
"f+" (function keys).

Too much manual work, besides Char(), Ctrl() and F(), the rest are a
long list to match with no tricks.
TODO: switch to using Nom for parsing
      https://github.com/Geal/nom

This commit brought to you by flight ZB7988