Reorganize vertical and horizontal scroll
Have each scroll focus on its direction, not mixed. There were some
horizontal moves in the vertical scroll because this function will
execute a scroll for sure. Meanwhile the horizontal one is meant to
check first, since it depends on the difference between focus, cursor
and line length.

To make this clearer also rename the functions to *do*_vertical_scroll()
and *consider*_horizontal_scroll().