Add Viewer.text_col to match cursor
There can be a difference between text column and cursor column if the
line has any characters that span more than one rune. Like a tab
occupies one column in the text, but TAB_SPACES columns in the viewer.
Use text_col element of Viewer to keep track of the real cursor
position.