Fix Viewer.do_forward_search() for tabs
The thing to have in mind here is that when we find a result, we find it
as a text position. We don't need to convert it to a cursor column because
set_cursor does it for us. set_cursor() uses the text position, since
internally it runs set_current_line() which does the match_cursor_text().

do_forward_search -> set_cursor -> set_current_line -> match_cursor_text